pyproject.toml
[project]
name = "resumescan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "VincentL01", email = "thang.luongcao@gmail.com" }
]
dependencies = [
    "google-genai>=1.26.0",
    "google-generativeai>=0.8.5",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.69",
    "langchain-ollama>=0.3.5",
    "langgraph>=0.5.3",
    "langgraph-checkpoint-sqlite>=2.0.10",
    "pdfplumber>=0.11.7",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
]
